requirements.txt
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
idna==2.10
oauthlib==3.1.0
python-dotenv==0.15.0
python-twitter==3.5
requests==2.25.1
requests-oauthlib==1.3.0
urllib3==1.26.2
